+++ src/sampling/cpu/index_select.hpp
@@ -6,7 +6,7 @@
 #include <cstring>
 
 template<typename DTYPE>
-DTYPE* IndexSelect(const DTYPE *feature, size_t dim, const std::vector<uint32_t> &index) {
+DTYPE* IndexSelect(const DTYPE *feature, uint64_t dim, const std::vector<uint32_t> &index) {
     DTYPE *ret = (DTYPE *) malloc(index.size() * dim * sizeof(DTYPE));
 #pragma omp parallel for
     for (uint32_t i = 0; i < index.size(); i++) {
